Updated das-apprenticeships to pull funding band maximum from approvals outer api, updated funding band maximum to correctly match int type and updated nuget package

+++ src/Domain/Apprenticeship/Approval.cs
@@ -16,9 +16,9 @@ namespace SFA.DAS.Apprenticeships.Domain.Apprenticeship
         public decimal AgreedPrice => _model.AgreedPrice;
         public long FunctionEmployerAccountId => _model.FundingEmployerAccountId;
         public FundingType FundingType => _model.FundingType;
-        public decimal FundingBandMaximum => _model.FundingBandMaximum;
+        public int FundingBandMaximum => _model.FundingBandMaximum;
 
-        internal static Approval New(long approvalsApprenticeshipId, long ukprn, long employerAccountId, string legalEntityName, DateTime? actualStartDate, DateTime? plannedEndDate, decimal agreedPrice, long fundingEmployerAccountId, FundingType fundingType, decimal fundingBandMaximum)
+        internal static Approval New(long approvalsApprenticeshipId, long ukprn, long employerAccountId, string legalEntityName, DateTime? actualStartDate, DateTime? plannedEndDate, decimal agreedPrice, long fundingEmployerAccountId, FundingType fundingType, int fundingBandMaximum)
         {
             return new Approval(new ApprovalModel
             {
