+++ src/Types/ApprenticeshipCreatedEvent.cs
@@ -14,5 +14,5 @@ public class ApprenticeshipCreatedEvent
     public string TrainingCode { get; set; }
     public long? FundingEmployerAccountId { get; set; }
     public FundingType FundingType { get; set; }
-    public decimal FundingBandMaximum { get; set; }
+    public int FundingBandMaximum { get; set; }
 }
